--- a/docs/src/assets/custom.css
+++ b/docs/src/assets/custom.css
@@ -1,4 +1,4 @@
-body {
+/* body {
     font-family: Helvetica;
     text-rendering: optimizeLegibility;
     line-height: 1.5;
@@ -51,4 +51,4 @@ a code {color: inherit;}
 
 pre code {color: black;}
 
-table code {color: black;}
+table code {color: black;} */