--- a/src/ScientificTypes.jl
+++ b/src/ScientificTypes.jl
@@ -7,6 +7,7 @@ export ColorImage, GrayImage
 export scitype, scitype_union, elscitype, coerce, coerce!, schema
 export info
 export autotype
+export MLJ
 
 # re-export from CategoricalArrays:
 export categorical