(I apologise for this big PR... but I think it's worth it, I would understand if it's not prioritized, the gist for working hard on this is that ST ends up being a fundamental op in MLJ and so, IMO, must be as optimized as possible)

This follows #77 and the discussion in #80. It is (mildly) **breaking** and so would require another minor release.

It brings **significant** performance improvements, the script ran is available in the `test/_benchmark.jl` script, condensed results below. In short, the performances are now on par with just calling `categorical` which is as good as we can make it.

```
@btime categorical($v1); # 28.495 ms (1000046 allocations: 34.34 MiB)
@btime coerce($v1, Multiclass);
#this branch ==> 28.755 ms (1000046 allocations: 34.34 MiB)
#master      ==> 46.380 ms (1000080 allocations: 31.48 MiB)

@btime categorical($v2); # 25.051 ms (799786 allocations: 28.22 MiB)
@btime coerce($v2, Union{Missing,Multiclass});
#this branch ==> 25.712 ms (799787 allocations: 28.22 MiB)
#master      ==> 139.152 ms (1599678 allocations: 51.69 MiB)

@btime coerce($df, autotype($df, :few_to_finite));
#this branch ==> 979.098 ms (11152394 allocations: 636.62 MiB)
#master      ==> 1.415 s (12273798 allocations: 609.21 MiB)
```

It doesn't affect `schema`. 

There are effectively two changes: a more careful treatment of missing values, leaving more responsibility to the user, this avoids unnecessarily checking whether a vector that has a Union{Missing,T} machine type really does have missing values. Second, a removal of compression which is explicitly discouraged in CategoricalArrays.

### Change  1 -- careful  treatment of missings

This is effectively  the result of the discussion at #80 . Basically it removes all checks for "real missing values". Two  important notes in changes of behaviour

- If the element machine type is `Union{Missing,T}`, the element scitype will be  `Union{Missing,S}` irrelevant of whether there are "true missings" or not
- An array of Any type coerced to some Finite type will have a `Union{Missing,<:Finite}` scitype irrelevant of whether it has missings or not

To clarify the second point, this is a behaviour that arises from `CategoricalArrays`: 

```julia
julia> a = categorical(Any[1,2,3]); eltype(a)
Union{Missing, CategoricalValue{Any,UInt32}}
```

### Change 2 -- no compression

The calls to `categorical` used to be like `categorical(v, true, ...)` where the  `true` corresponds to the `compress` argument. There may be a good reason to do  this but I don't immediately see it; the gain in memory is irrelevant especially as the previous behaviour required to have copies of the original vector by broadcasting over it. Also the use of that argument is [explicitly discouraged](https://juliadata.github.io/CategoricalArrays.jl/latest/using.html#CategoricalArrays.compress) due to performance issues resulting from type  instabilities. 

I may have overlooked some important reason why this is useful but in that case I suggest we reexport `CategoricalArrays.compress` and  let people who know what they're doing use it.

### Tightening

There may now be _more_ cases where a user would get scitype `Union{Missing,S}` where there are actually no missings and this could be a nuisance to users. I don't think this will arise very frequently but even infrequently I could see this causing confusion to a user. 
So I added a `tight` keyword to `coerce` (it can be passed to `autotype`) which would do what we were previously doing. Of course the use of it will go with a tax on performance.

```julia
julia> a = categorical(Any[1,2,3]); c = coerce(a, Multiclass); elscitype(c)
 Warning: Trying to coerce from `Any` to `Multiclass` with categoricals.
│ Coerced to Union{Missing,Multiclass} instead.
└ @ ScientificTypes ~/.julia/dev/ScientificTypes/src/conventions/mlj/utils.jl:4
Union{Missing, Multiclass{3}}

julia> a = categorical(Any[1,2,3]); c = coerce(a, Multiclass; tight=true); elscitype(c)
Multiclass{3}
```