@ablaom 2020-01-07
@tlienart Thanks for the work on this PR. I am sorry I cannot give a more postive review at this point.

According to the definition in the documents, which until now has been inviolate, *the scitype of an AbstractArray is AbstractArray{U} where U is the union of the element scitypes*. Of course, implementations do not literally need to compute element scitype unions, but this is the behaviour prescribed in the current design.

So if an array has no missing elements, then its scitype can *never* be `<:AbstractArray{Union{Missing,S}, N} where {S,N}`. As far as I am aware this definition is nowhere violated in the existing implementation. But in the proposed PR we already have (before considering categoricals or `Any` types):

```julia
scitype([1, 2, missing][1:2])
AbstractArray{Union{Missing, Count},1}
```

instead of 
```julia
julia> scitype([1, 2, missing][1:2])
AbstractArray{Count,1}
```

consistent with the definition.

A performance improvement is not enough to justify the proposed change. One would need to propose a new (presumably more complicated) definition and justify that first. (If categorical arrays had included `ordered` and `n_classes` as type parameters, then we could have defined scitypes as a simple map from ordinary types, rather than operating at the level of instances, but they didn't.)

If we accept the existing definition as inviolate, then to compute the scitype of any vector `v` for which `Missing <: eltype(v)` one cannot avoid actually parsing the array for missing values, as far as I can see.