@tlienart 2020-01-08
I understand your point but I don't think it's reasonable to ask the user to pay that price **all the times** when, in fact, the case would happen rarely.

I'd like to think more about this but here's an additional thought:

Let's say the user has a dataset X with (true) missings, then the user takes a subset of that dataset without true missings, trains a model on it. Then tries to predict on a chunk that has missings; the scitype would be different. I think this is an issue.

I would be in favour of adding an explicit keyword for `scitype` in the same way I added one for `coerce`. Also I feel the behaviour is restricted (`Union{Missing,T}` with no true missing and `Any`) and can be explained well. 

We can discuss this more but I think we really want to avoid ScientificTypes adding unnecessary overhead which users then assume is caused by "machine" and then generally assume that MLJ makes stuff slow. The type of performance gains indicated above should be considered in the light of "real-world" datasets where you would multiply this by the number of features which may be very large.