@ablaom 2020-01-08
What about the following alternative definition:

---

*The scitype of an `AbstractArray`, `A`, is always`AbstractArray{U}` where `U` is the union of the element scitypes,  with one exception: If `typeof(A) <: AbstractArray{Union{Missing,T}}` for some `T`, then the scitype is `AbstractArray{Union{Missing, U}}`, **even if `A` has no missing values**.* 

This exception is made for performance reasons. If one want's to override it, one uses `scitype(A, tight=true)`. 

---

Could we adhere strictly to this?