@tlienart 2020-01-10
Thanks for being flexible on this Anthony, your definition works fine with me but there is still the corner case of the `Any` type. 

For that one though, I think the story is less clear and I would be happy to  go  with the "broadcast by default" which would correspond to the behaviour that was there initially. For the sake of clarifying the discussion consider:

```julia
julia> using CategoricalArrays
julia> a = Any[1,2,3];
julia> ca = categorical(a)
3-element CategoricalArray{Any,1,UInt32}:
 1
 2
 3

julia> eltype(ca)
Union{Missing, CategoricalValue{Any,UInt32}}
```

What the current PR is doing:

1. AbstractArray{Any} => Continuous : broadcast `float` (warn if hasmissing that coercing to Union{Missing,Continuous})
1. AbstractArray{Any} => Count: broadcast `int` (warn if hasmissing that coercing to Union{Missing,Count})
1. AbstractArray{Any} => Multiclass or OrderedFactor : just call `categorical` (warn that coercing to Union{Missing, Multiclass} or Union{Missing, OrderedFactor})

Then

1. CategoricalArray{Any} => Count or Continuous: same story as with AbstractArray
1. CategoricalArray{Any} => Multiclass or OrderedFactor : just return the vector (possibly after ordering levels) with warning that coercing to Union{Missing,Multiclass} or Union{Missing,OrderedFactor}

Based on your input I'll adjust the current PR to take into account the new definition and have explicit tests for all those cases.