@ablaom 2020-01-12
Yes the coerce behaviour for `Any` sounds good. I think it's important to document it in the docstrings, if only to remind ourselves what that behaviour is. 

Please update the new  definition of scitype in /docs. You can just copy and paste what I wrote above.