@tlienart 2020-01-13
@ablaom, all done. Details of further changes below:

* docs adjustment based on your comment
* added a `tight` keyword to `scitype` and to  `schema` (on top of `coerce`)
* some code cleanup to not have to use the  `Val(Scitype(T, C))` trick, this makes calls simpler, the code more readable and reduces the need for specialised methods
* added a bunch of additional tests for the `tight` keyword as well as for the specific conversion cases with `Any` type discussed earlier
* updated Documenter to work with documenter 0.23 and 0.24